Extract elevator handling
