+++ src/main/java/org/opentripplanner/graph_builder/module/osm/OpenStreetMapModule.java
@@ -4,14 +4,11 @@ import com.google.common.collect.Iterables;
 import gnu.trove.iterator.TLongIterator;
 import gnu.trove.list.TLongList;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.OptionalInt;
 import java.util.Set;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Geometry;
@@ -22,7 +19,6 @@ import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.framework.i18n.NonLocalizedString;
 import org.opentripplanner.framework.logging.ProgressTracker;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueStore;
-import org.opentripplanner.graph_builder.issue.api.Issue;
 import org.opentripplanner.graph_builder.issues.Graphwide;
 import org.opentripplanner.graph_builder.issues.StreetCarSpeedZero;
 import org.opentripplanner.graph_builder.issues.TurnRestrictionBad;
@@ -44,22 +40,15 @@ import org.opentripplanner.street.model.TurnRestriction;
 import org.opentripplanner.street.model.edge.AreaEdge;
 import org.opentripplanner.street.model.edge.AreaEdgeList;
 import org.opentripplanner.street.model.edge.Edge;
-import org.opentripplanner.street.model.edge.ElevatorAlightEdge;
-import org.opentripplanner.street.model.edge.ElevatorBoardEdge;
-import org.opentripplanner.street.model.edge.ElevatorHopEdge;
-import org.opentripplanner.street.model.edge.FreeEdge;
 import org.opentripplanner.street.model.edge.NamedArea;
 import org.opentripplanner.street.model.edge.StreetEdge;
 import org.opentripplanner.street.model.note.StreetNoteAndMatcher;
 import org.opentripplanner.street.model.vertex.BarrierVertex;
-import org.opentripplanner.street.model.vertex.ElevatorOffboardVertex;
-import org.opentripplanner.street.model.vertex.ElevatorOnboardVertex;
 import org.opentripplanner.street.model.vertex.ExitVertex;
 import org.opentripplanner.street.model.vertex.IntersectionVertex;
 import org.opentripplanner.street.model.vertex.OsmBoardingLocationVertex;
 import org.opentripplanner.street.model.vertex.OsmVertex;
 import org.opentripplanner.street.model.vertex.Vertex;
-import org.opentripplanner.transit.model.basic.Accessibility;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -170,14 +159,7 @@ public class OpenStreetMapModule implements GraphBuilderModule {
     return elevationData;
   }
 
-  private Issue invalidDuration(OSMWithTags element, String v) {
-    return Issue.issue(
-      "InvalidDuration",
-      "Duration for osm node %d is not a number: '%s'; it's replaced with '-1' (unknown).",
-      element.getId(),
-      v
-    );
-  }
+  private record StreetEdgePair(StreetEdge main, StreetEdge back) {}
 
   protected class Handler {
 
@@ -190,7 +172,7 @@ public class OpenStreetMapModule implements GraphBuilderModule {
     private final OSMDatabase osmdb;
     // track OSM nodes which are decomposed into multiple graph vertices because they are
     // elevators. later they will be iterated over to build ElevatorEdges between them.
-    private final HashMap<Long, HashMap<OSMLevel, OsmVertex>> multiLevelNodes = new HashMap<>();
+    private final HashMap<Long, Map<OSMLevel, OsmVertex>> multiLevelNodes = new HashMap<>();
     // track OSM nodes that will become graph vertices because they appear in multiple OSM ways
     private final Map<Long, IntersectionVertex> intersectionNodes = new HashMap<>();
     /**
@@ -256,7 +238,13 @@ public class OpenStreetMapModule implements GraphBuilderModule {
         graph.getVehicleParkingService().updateVehicleParking(parkingLots, List.of());
       }
 
-      buildElevatorEdges(graph);
+      var elevatorProcessor = new ElevatorProcessor(
+        issueStore,
+        osmdb,
+        multiLevelNodes,
+        intersectionNodes
+      );
+      elevatorProcessor.buildElevatorEdges(graph);
 
       unifyTurnRestrictions();
 
@@ -365,7 +353,7 @@ public class OpenStreetMapModule implements GraphBuilderModule {
       iv = intersectionNodes.get(nid);
       if (iv == null) {
         Coordinate coordinate = getCoordinate(node);
-        String label = getNodeLabel(node);
+        String label = String.format(nodeLabelFormat, node.getId());
         String highway = node.getTag("highway");
         if ("motorway_junction".equals(highway)) {
           String ref = node.getTag("ref");
@@ -422,8 +410,16 @@ public class OpenStreetMapModule implements GraphBuilderModule {
       return iv;
     }
 
-    private OptionalInt parseDuration(OSMWithTags element) {
-      return element.getTagAsInt("duration", v -> issueStore.add(invalidDuration(element, v)));
+    /**
+     * Returns the length of the geometry in meters.
+     */
+    private static double getGeometryLengthMeters(Geometry geometry) {
+      Coordinate[] coordinates = geometry.getCoordinates();
+      double d = 0;
+      for (int i = 1; i < coordinates.length; ++i) {
+        d += SphericalDistanceLibrary.distance(coordinates[i - 1], coordinates[i]);
+      }
+      return d;
     }
 
     private List<AreaGroup> groupAreas(Collection<Area> areas) {
@@ -652,179 +648,6 @@ public class OpenStreetMapModule implements GraphBuilderModule {
       }
     }
 
-    private void buildElevatorEdges(Graph graph) {
-      /* build elevator edges */
-      for (Long nodeId : multiLevelNodes.keySet()) {
-        OSMNode node = osmdb.getNode(nodeId);
-        // this allows skipping levels, e.g., an elevator that stops
-        // at floor 0, 2, 3, and 5.
-        // Converting to an Array allows us to
-        // subscript it so we can loop over it in twos. Assumedly, it will stay
-        // sorted when we convert it to an Array.
-        // The objects are Integers, but toArray returns Object[]
-        HashMap<OSMLevel, OsmVertex> vertices = multiLevelNodes.get(nodeId);
-
-        /*
-         * first, build FreeEdges to disconnect from the graph, GenericVertices to serve as attachment points, and ElevatorBoard and
-         * ElevatorAlight edges to connect future ElevatorHop edges to. After this iteration, graph will look like (side view): +==+~~X
-         *
-         * +==+~~X
-         *
-         * +==+~~X
-         *
-         * + GenericVertex, X EndpointVertex, ~~ FreeEdge, == ElevatorBoardEdge/ElevatorAlightEdge Another loop will fill in the
-         * ElevatorHopEdges.
-         */
-        OSMLevel[] levels = vertices.keySet().toArray(new OSMLevel[0]);
-        Arrays.sort(levels);
-        ArrayList<Vertex> onboardVertices = new ArrayList<>();
-        for (OSMLevel level : levels) {
-          // get the node to build the elevator out from
-          OsmVertex sourceVertex = vertices.get(level);
-          String sourceVertexLabel = sourceVertex.getLabel();
-          String levelName = level.longName;
-
-          createElevatorVertices(
-            graph,
-            onboardVertices,
-            sourceVertex,
-            sourceVertexLabel,
-            levelName
-          );
-        }
-        int travelTime = parseDuration(node).orElse(-1);
-
-        var wheelchair = node.getWheelchairAccessibility();
-
-        createElevatorHopEdges(
-          onboardVertices,
-          wheelchair,
-          node.isTagTrue("bicycle"),
-          levels.length,
-          travelTime
-        );
-      } // END elevator edge loop
-
-      // Add highway=elevators to graph as elevators
-      Iterator<OSMWay> elevators = osmdb.getWays().stream().filter(this::isElevatorWay).iterator();
-
-      while (elevators.hasNext()) {
-        OSMWay elevatorWay = elevators.next();
-
-        List<Long> nodes = Arrays
-          .stream(elevatorWay.getNodeRefs().toArray())
-          .filter(nodeRef ->
-            intersectionNodes.containsKey(nodeRef) && intersectionNodes.get(nodeRef) != null
-          )
-          .boxed()
-          .toList();
-
-        ArrayList<Vertex> onboardVertices = new ArrayList<>();
-        for (int i = 0; i < nodes.size(); i++) {
-          Long node = nodes.get(i);
-          var sourceVertex = intersectionNodes.get(node);
-          String sourceVertexLabel = sourceVertex.getLabel();
-          String levelName = elevatorWay.getId() + " / " + i;
-          createElevatorVertices(
-            graph,
-            onboardVertices,
-            sourceVertex,
-            elevatorWay.getId() + "_" + sourceVertexLabel,
-            levelName
-          );
-        }
-
-        int travelTime = parseDuration(elevatorWay).orElse(-1);
-        int levels = nodes.size();
-        var wheelchair = elevatorWay.getWheelchairAccessibility();
-
-        createElevatorHopEdges(
-          onboardVertices,
-          wheelchair,
-          elevatorWay.isTagTrue("bicycle"),
-          levels,
-          travelTime
-        );
-        LOG.debug("Created elevatorHopEdges for way {}", elevatorWay.getId());
-      }
-    }
-
-    private boolean isElevatorWay(OSMWay way) {
-      if (!way.hasTag("highway")) {
-        return false;
-      }
-      if (!"elevator".equals(way.getTag("highway"))) {
-        return false;
-      }
-      if (osmdb.isAreaWay(way.getId())) {
-        return false;
-      }
-
-      TLongList nodeRefs = way.getNodeRefs();
-      // A way whose first and last node are the same is probably an area, skip that.
-      // https://www.openstreetmap.org/way/503412863
-      // https://www.openstreetmap.org/way/187719215
-      return nodeRefs.get(0) != nodeRefs.get(nodeRefs.size() - 1);
-    }
-
-    private void createElevatorVertices(
-      Graph graph,
-      ArrayList<Vertex> onboardVertices,
-      IntersectionVertex sourceVertex,
-      String sourceVertexLabel,
-      String levelName
-    ) {
-      ElevatorOffboardVertex offboardVertex = new ElevatorOffboardVertex(
-        graph,
-        sourceVertexLabel + "_offboard",
-        sourceVertex.getX(),
-        sourceVertex.getY(),
-        new NonLocalizedString(levelName)
-      );
-
-      new FreeEdge(sourceVertex, offboardVertex);
-      new FreeEdge(offboardVertex, sourceVertex);
-
-      ElevatorOnboardVertex onboardVertex = new ElevatorOnboardVertex(
-        graph,
-        sourceVertexLabel + "_onboard",
-        sourceVertex.getX(),
-        sourceVertex.getY(),
-        new NonLocalizedString(levelName)
-      );
-
-      new ElevatorBoardEdge(offboardVertex, onboardVertex);
-      new ElevatorAlightEdge(onboardVertex, offboardVertex, new NonLocalizedString(levelName));
-
-      // accumulate onboard vertices to so they can be connected by hop edges later
-      onboardVertices.add(onboardVertex);
-    }
-
-    private void createElevatorHopEdges(
-      ArrayList<Vertex> onboardVertices,
-      Accessibility wheelchair,
-      boolean bicycleAllowed,
-      int levels,
-      int travelTime
-    ) {
-      // -1 because we loop over onboardVertices two at a time
-      for (int i = 0, vSize = onboardVertices.size() - 1; i < vSize; i++) {
-        Vertex from = onboardVertices.get(i);
-        Vertex to = onboardVertices.get(i + 1);
-
-        // default permissions: pedestrian, wheelchair, check tag bicycle=yes
-        StreetTraversalPermission permission = bicycleAllowed
-          ? StreetTraversalPermission.PEDESTRIAN_AND_BICYCLE
-          : StreetTraversalPermission.PEDESTRIAN;
-
-        if (travelTime > -1 && levels > 0) {
-          ElevatorHopEdge.bidirectional(from, to, permission, wheelchair, levels, travelTime);
-        } else {
-          ElevatorHopEdge.bidirectional(from, to, permission, wheelchair);
-        }
-      }
-    }
-
     private void unifyTurnRestrictions() {
       // Note that usually when the from or to way is not found, it's because OTP has already
       // filtered that way. So many missing edges are not really problems worth issuing warnings on.
@@ -1049,26 +872,6 @@ public class OpenStreetMapModule implements GraphBuilderModule {
       return new Coordinate(osmNode.lon, osmNode.lat);
     }
 
-    private String getNodeLabel(OSMNode node) {
-      return String.format(nodeLabelFormat, node.getId());
-    }
-
-    private String getLevelNodeLabel(OSMNode node, OSMLevel level) {
-      return String.format(levelnodeLabelFormat, node.getId(), level.shortName);
-    }
-
-    /**
-     * Returns the length of the geometry in meters.
-     */
-    private double getGeometryLengthMeters(Geometry geometry) {
-      Coordinate[] coordinates = geometry.getCoordinates();
-      double d = 0;
-      for (int i = 1; i < coordinates.length; ++i) {
-        d += SphericalDistanceLibrary.distance(coordinates[i - 1], coordinates[i]);
-      }
-      return d;
-    }
-
     /**
      * Handle oneway streets, cycleways, and other per-mode and universal access controls. See
      * http://wiki.openstreetmap.org/wiki/Bicycle for various scenarios, along with
@@ -1091,7 +894,7 @@ public class OpenStreetMapModule implements GraphBuilderModule {
 
       LineString backGeometry = geometry.reverse();
       StreetEdge street = null, backStreet = null;
-      double length = this.getGeometryLengthMeters(geometry);
+      double length = getGeometryLengthMeters(geometry);
 
       var permissionPair = OSMFilter.getPermissions(permissions, way);
       var permissionsFront = permissionPair.main();
@@ -1203,7 +1006,7 @@ public class OpenStreetMapModule implements GraphBuilderModule {
      */
     private OsmVertex recordLevel(OSMNode node, OSMWithTags way) {
       OSMLevel level = osmdb.getLevelForWay(way);
-      HashMap<OSMLevel, OsmVertex> vertices;
+      Map<OSMLevel, OsmVertex> vertices;
       long nodeId = node.getId();
       if (multiLevelNodes.containsKey(nodeId)) {
         vertices = multiLevelNodes.get(nodeId);
@@ -1213,7 +1016,7 @@ public class OpenStreetMapModule implements GraphBuilderModule {
       }
       if (!vertices.containsKey(level)) {
         Coordinate coordinate = getCoordinate(node);
-        String label = this.getLevelNodeLabel(node, level);
+        String label = String.format(levelnodeLabelFormat, node.getId(), level.shortName);
         OsmVertex vertex = new OsmVertex(
           graph,
           label,
@@ -1229,6 +1032,4 @@ public class OpenStreetMapModule implements GraphBuilderModule {
       return vertices.get(level);
     }
   }
-
-  private record StreetEdgePair(StreetEdge main, StreetEdge back) {}
 }
